Added open/close MR feature
The client can now change its MR status on the server! I added a function in the server file to search for a client in the main client list based off of its pthread_id and change the status of that client. The entire implementation of the actual client/server communication was just a handful of lines. It may be a teeny tiny change but I'm trying to stay on top of documentation and this whole github thing.